Keep precommit helpers out of runner coverage
