+++ test/precommit/environment.js
@@ -1,0 +1,26 @@
+const GIT_LOCAL_ENV_VARS = [
+  "GIT_ALTERNATE_OBJECT_DIRECTORIES",
+  "GIT_COMMON_DIR",
+  "GIT_CONFIG",
+  "GIT_CONFIG_COUNT",
+  "GIT_CONFIG_PARAMETERS",
+  "GIT_DIR",
+  "GIT_GRAFT_FILE",
+  "GIT_IMPLICIT_WORK_TREE",
+  "GIT_INDEX_FILE",
+  "GIT_NO_REPLACE_OBJECTS",
+  "GIT_OBJECT_DIRECTORY",
+  "GIT_PREFIX",
+  "GIT_REPLACE_REF_BASE",
+  "GIT_SHALLOW_FILE",
+  "GIT_WORK_TREE",
+];
+
+export const getStepEnvironment = (environment, verbose) => ({
+  ...Object.fromEntries(
+    Object.entries(environment).filter(
+      ([name]) => !GIT_LOCAL_ENV_VARS.includes(name),
+    ),
+  ),
+  VERBOSE: verbose ? "1" : "0",
+});
