+++ test/precommit/runner.js
@@ -19,6 +19,7 @@ import {
   write,
   yellow,
 } from "#test/precommit/colors.js";
+import { getStepEnvironment } from "#test/precommit/environment.js";
 import { getMergeConflictWarning } from "#test/precommit/merge-warning.js";
 import { promptToPushCheckedInChanges } from "#test/precommit/push.js";
 import { getSteps } from "#test/precommit/steps.js";
@@ -33,34 +34,7 @@ import {
   reportCoverageFailures,
 } from "#test/test-runner-utils.js";
 
-const GIT_LOCAL_ENV_VARS = [
-  "GIT_ALTERNATE_OBJECT_DIRECTORIES",
-  "GIT_COMMON_DIR",
-  "GIT_CONFIG",
-  "GIT_CONFIG_COUNT",
-  "GIT_CONFIG_PARAMETERS",
-  "GIT_DIR",
-  "GIT_GRAFT_FILE",
-  "GIT_IMPLICIT_WORK_TREE",
-  "GIT_INDEX_FILE",
-  "GIT_NO_REPLACE_OBJECTS",
-  "GIT_OBJECT_DIRECTORY",
-  "GIT_PREFIX",
-  "GIT_REPLACE_REF_BASE",
-  "GIT_SHALLOW_FILE",
-  "GIT_WORK_TREE",
-];
-
 const isVerbose = () => process.argv.includes("--verbose");
-
-export const getStepEnvironment = (environment, verbose) => ({
-  ...Object.fromEntries(
-    Object.entries(environment).filter(
-      ([name]) => !GIT_LOCAL_ENV_VARS.includes(name),
-    ),
-  ),
-  VERBOSE: verbose ? "1" : "0",
-});
 
 const finishStep = ({ step, stdout, stderr, status, elapsed }) => {
   const mark = status === 0 ? green("✓") : red("✗");
